Added some logos AND touchable buttons to easily navigate while testing our code.
